add: implement Tea class with validation and description methods
