+++ courses/backend/advanced-javascript/week4-assigment/exercise1.js
@@ -1,69 +1,58 @@
-import { teas } from "../teas.js";
-export class Tea {
-    constructor(name, type, origin, pricePerGram, organic) {
-        // Validate:
-        // - name must be a non-empty string
-        if (typeof name !== "string" || name.trim() === "") {
-            throw new Error("Name is required");
-        }
-        // - type must be one of: "green", "black", "herbal", "oolong", "white"
-        const validTypes = ["green", "black", "herbal", "oolong", "white"];
-        if (!validTypes.includes(type)) {
-            throw new Error(`Invalid type: ${type}`);
-        }
-        // - pricePerGram must be a positive number
-        if (typeof pricePerGram !== "number" || pricePerGram <= 0) {
-            throw new Error("Price per gram must be positive");
-        }
-        // Store all properties
-        this.name = name;
-        this.type = type;
-        this.origin = origin;
-        this.pricePerGram = pricePerGram;
-        this.organic = organic;
+import { teas } from "./teas.js";
+class Inventory {
+    constructor() {
+
+        // Store a Map or object of tea ID → { tea, stockCount }
     }
 
-    priceFor(grams) {
-        return this.pricePerGram * grams;
+    add(tea, stockCount) {
+        // Add a tea to inventory
     }
 
-    describe() {
-        const pricePer100g = (this.pricePerGram * 100).toFixed(2);
-        let desc = `${this.name} (${this.type}) from ${this.origin} - ${pricePer100g} DKK/100g`;
-        if (this.organic) {
-            desc += " [organic]";
-        }
-        return desc;
+    sell(teaName, grams) {
+        // Reduce stock. Throw if not enough stock.
     }
 
-    static fromObject(obj) {
-        return new Tea(
-            obj.name,
-            obj.type,
-            obj.origin,
-            obj.pricePerGram,
-            obj.organic
-        );
+    restock(teaName, grams) {
+        // Increase stock
+    }
+
+    getStock(teaName) {
+        // Return current stock count for a tea
+    }
+
+    getLowStock(threshold) {
+        // Return array of { tea, stockCount } where stock < threshold
+        // Use .filter()
+    }
+
+    getTotalValue() {
+        // Sum of (pricePerGram * stockCount) for all items
+        // Use .reduce()
     }
 }
 
-// Test validation:
-try {
-    new Tea("", "green", "Japan", 0.12, true);
-} catch (e) {
-    console.log(e.message);
-} // "Name is required"
-
-try {
-    new Tea("Test", "purple", "Japan", 0.12, true);
-} catch (e) {
-    console.log(e.message);
-} // "Invalid type: purple"
-
-// Test factory method:
+// Test:
 const teaInstances = teas.map(Tea.fromObject);
-console.log(teaInstances.length); // 20
-console.log(teaInstances[0].describe());
-// "Sencha (green) from Japan - 12.00 DKK/100g [organic]"
-console.log(teaInstances[1].describe());
-// "Earl Grey (black) from India - 8.00 DKK/100g"
+const inventory = new Inventory();
+
+teaInstances.forEach((tea) => {
+    const data = teas.find((t) => t.name === tea.name);
+    inventory.add(tea, data.stockCount);
+});
+
+console.log("Sencha stock:", inventory.getStock("Sencha")); // 150
+
+inventory.sell("Sencha", 50);
+console.log("After selling 50g:", inventory.getStock("Sencha")); // 100
+
+console.log("Low stock (< 50):");
+inventory.getLowStock(50).forEach((item) => {
+    console.log(`- ${item.tea.name}: ${item.stockCount}g`);
+});
+
+console.log(
+    "Total inventory value:",
+    inventory.getTotalValue().toFixed(2),
+    "DKK",
+);
